Rename examples/bert -> examples/bert_pretraining
