+++ examples/bert_pretraining/bert_config.py
@@ -77,9 +77,3 @@ TRAINING_CONFIG = {
     # Percentage of training steps used for learning rate warmup.
     "warmup_percentage": 0.1,
 }
-
-FINETUNING_CONFIG = {
-    "batch_size": 32,
-    "epochs": 3,
-    "learning_rates": [5e-5, 4e-5, 3e-5, 2e-5],
-}
